requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
openai>=1.50.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
jinja2>=3.1.4
python-multipart>=0.0.12
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
langfuse>=2.0.0
anthropic>=0.50.0
posthog>=3.0.0
boto3>=1.35.0
itsdangerous>=2.1.0
